5.14. Модуляция + чистка кодв
